Add client parameter to READ-COMMON, move critical code out of read-cst.lisp
READ-COMMON now takes a client object as the first argument. This is a
first step towards additional customizability.

It also enables the separation of critical and cst-specific parts in
the READ-COMMON :around method previously defined in
read-cst.lisp.

More specifically, binding of the *BACKQUOTE[-IN-SUBFORMS]-ALLOWED-P*
variables is now done in code of the eclector system, independent of
whether eclector-concrete-syntax-tree is loaded or not. This also
resolves the issue of referring to these variables with double-colon
package qualification in read-cst.lisp.

No functional change intended.