Quasiquote and unquote are allowed by default
Quasiquote and unquote are now opt-out instead of opt-in. This allows
quasiquotation in custom reader macros by default. The new macro
WITH-FORBIDDEN-QUASIQUOTATION is used by Eclector (and can be used in
custom reader macros) to controls this behavior.

As a further improvement, the new
variables *{QUASIQUOTE,UNQUOTE}-FORBIDDEN* can be bound to non-NIL
symbols to indicate the context in which the respective quasiquotation
construct has been disabled. These context names are turned into
translatable strings using the generic function CONTEXT-NAME and
reported in error messages.

fixes #57