LEFT-PARENTHESIS always reads until #\), not some opposite character
The #\( reader macro is specified to read until the next #\)
character, not some "opposite" character of the character the reader
macro function was invoked with.

This came up because ANSI tests install LEFT-PARENTHESIS as the reader
macro for e.g. #\a and expect (read-from-string "a)") => (), that is
the corresponding closing list delimiter is always #\).