Implement and default to relaxed syntax for SHARPSIGN-C
Depending on the new ALLOW-NON-LIST parameter, SHARPSIGN-C can now
read the characters following "#C" either via READ or via
%READ-MAYBE-NOTHING. The former allows, for example, read-time
evaluation and skippable input such as whitespace, block comments or
reader conditional preceding the and in the list of complex parts. The
latter requires the list of complex parts to immediately follow the
"#C" and use list syntax, that is, only "#C(" is accepted.

By default, ALLOW-NON-LIST is true, resulting in the less strict of
the two possible behaviors.

The differences has deep, and maybe unfortunate, implications for
parse results: If ALLOW-NON-LIST is true and thus recursive READ is
used, the parse result tree for the complex literal will be

  node for complex literal
  - node for list of complex parts
    - node for first complex part
    - node for second complex part

On other hand, if ALLOW-NON-LIST is false, the list of complex parts
will not be represented explicitly:

  node for complex literal
  - node for first complex part
  - node for second complex part

fixes #70