Move *CLIENT* and source location protocol to base module
This change moves the *CLIENT* special variable and the source
location protocol which consists of the SOURCE-POSITION and
MAKE-SOURCE-RANGE generic functions into the base module. This move
enables code in the reader module to use the location protocol. In
particular code in the reader module that signals errors can now call
SOURCE-POSITION and/or MAKE-SOURCE-RANGE and attach the resulting
objects to the signaled conditions.

The compatibility infrastructure in the parse-result module is
intended to be removed after a grace period.