Use SOURCE-POSITION for locations of signaled errors
This way clients can customize the representation of source locations
in signaled conditions - in addition to parse results for which the
customization was already possible before this change. Since Eclector
does not access or process the source locations in signaled
conditions, clients have complete freedom with respect to the chosen
representation.