Control loading of code/reader/variables.lisp via a new feature
The feature :ECLECTOR-DEFINE-CL-VARIABLES controls whether
code/reader/variables.lisp is loaded.

When Eclector is loaded into a complete Common Lisp system, the file
usually must not be loaded since it would redefine variables which are
already present in the Common Lisp system.

However, if Eclector is loaded in order to be used as the reader of a
Common Lisp system, loading the file is one way to define the standard
variables which control the Common Lisp reader.