FIND-CHARACTER is also called in the #\<single character> case
This change allows clients to observe the creation of all character
literals. Clients could also replace a character literal spelled
as #\<single-character> with something else.

The new behavior is compatible with the previous behavior since the
FIND-CHARACTER method which is specialized to the character name being
a CHARACTER just returns that character.