New protocol for sharpsign {equals,sharpsign} and fixup
The main idea of this change is to use recognizable placeholder
objects within read objects instead of the MAPPING parameter of the
FIXUP function. These "labeled objects" form the basis of the new
protocol.

Another aspect of this change is the handling of FIXUP calls: instead
of calling FIXUP on the entire read object, call FIXUP after reading a
circular object.

The helper macros FIXUP-{CASE,PLACE-USING-VALUE,PLACE} are exported
for use in multiple Eclector modules but are not yet intended to be
used by client code.

Thanks to Alex Wood for testing the code changes, finding a big
oversight and reviewing the documentation.