Build labeled object tree to minimize fixup work
Before this change, cases like

  #1=(1 #1# #2=(2 #2# ... #100=(100 #100#)))

were not processed efficiently since FIXUP would be called for each of
the nested labeled objects, repeatedly traversing the portion of the
object graph that is reachable from the respective object.

The above problem could be solved with a simple stack of labeled
objects which are currently being processed. However, cases like

  #1=(1 #2=(2 ... #2#) ... #1#)

did not work well because after reading #2=(2 ... #2#), it was not yet
clear that the outer object would also need fixing up, so the object
arising for #2=(2 ... #2#) was fixed up immediately. Later, after
finishing reading the outer object, the whole tree was fixed up
again. So quadratic behavior was not avoided in this case.

To address both cases, this change introduces a tree representation for
labeled objects that need some kind of fixup work. Fixup work for the
labeled object that is associated with a node in this tree can be
delayed if an ancestor node also requires fixup work since the
traversal and fixup for the ancestor labeled object will include the
descendant labeled object.