Perform fixup in parse results
Before this change, something like

  (eclector.concrete-syntax-tree:read-from-string
   "#1=(#1#)")

produced a CST object, say cst, which failed to satisfy

  (eq (cst:first cst)       cst)
  (eq (cst:raw (first cst)) (cst:raw cst))

The properties now hold.

With this change, for an input of the form #1=(#1#), the default
behavior is to create a parse result for the READ call which
reads (#1#) but not for the READ call which reads the whole input
including the label definition #1=(#1#). Similarly, by default, the
object referenced by #1# is represented as a parse result (the same
parse result as the one associated with the definition), but not the
reference itself. This behavior can be changed by defining methods on
ECLECTOR.PARSE-RESULT:MAKE-EXPRESSION-RESULT which are specialized to

      (eql eclector.parse-result:**definition**)
  and (eql eclector.parse-result:**reference**)

respectively.