CST classes and client mixins for explicit defs and refs
By default, labeled object definitions and references are not
represented as distinct CST nodes. This change introduces the mixin
classes DEFINITION-CSTS-MIXIN and REFERENCE-CSTS-MIXIN which can be
mixed into client classes to change this behavior. A client class that
is a subclass of DEFINITION-CSTS-MIXIN will create DEFINITION-CST
instances which represent labeled object definitions. Similarly for
REFERENCE-CSTS-MIXIN, REFERENCE-CST and labeled object references.