New protocol for querying and binding reader state aspects
The protocol allows clients as well as the reader to query and bind
reader state aspects which change the behavior of the reader such as
the current package, the current readtable and the current read base.

Clients can use this protocol to control the reader state in other
ways than binding the Common Lisp variables, for example by storing
the values of reader state aspects in context objects.

Furthermore, implementations which use Eclector as the Common Lisp
reader can use this protocol to tie the CL:*READTABLE* aspect to the
CL:*READTABLE* variable instead of the ECLECTOR.READER:*READTABLE*
variable.