tests: Bind *PRINT-CIRCLE* to T when PRINTing circular structures
The affected tests use GEN-LABELS-AND-REFERENCES to build random,
possibly circular structures and turn those into strings using
PRIN1-TO-STRING. This print operation fails when *PRINT-CIRCLE* is
false at the time the respective test is run. *PRINT-CIRCLE* can be
false when the respective test is run either directly or through
FIVEAM:RUN!. When one of the RUN-TESTS functions or ASDF:TEST-OP is
used, the test is run with *PRINT-CIRCLE* bound to T. The latter
behavior is the reason why the problem was not detected by CI or runs
of the whole test suite in multiple implementations.

One way to illustrate the problem is to load the eclector/test system
and evaluate (fiveam:run! :eclector.reader).

fixes #77