Do not use GENSYM at read-time in the Eclector code
Primarily so that the names of symbols in READ expressions do not
depend on the initial value of *GENSYM-COUNTER* (and changes to the
value the implementation might make in the process of READing). Before
this change, the READ-CODE/HOST-EQUIVALENCE test would fail on
implementations that change *GENSYM-COUNTER* in the process of READing
since CL:READ and ECLECTOR.READER:READ would produce symbols with
different (that is not STRING=) names for source code fragments of the
form #.(gensym ...).

For READ-CHAR{,-OR-ERROR,-OR-RECOVERABLE-ERROR} and PEEK-CHAR, using
the input stream as the sentinel value is simpler.

For %SHARPSIGN-C and %SHARPSIGN-S, use MAKE-SYMBOL at read-time
instead of GENSYM.

In the READ-CODE/HOST-EQUIVALENCE test, do not bind *GENSYM-COUNTER*
since there should be no GENSYM calls at read-time.

fixes #78