Allow parse results for labeled objects to contain children
Before this change, clients could not build a complete concrete syntax
tree for input of the form #1=#|foo|#34, that is a labeled object
definition which lexically contains skipped material. Parse results
were constructed for the skipped material but could not be included in
the labeled object definition parse result. The call which created the
labeled object definition parse result was

  (make-expression-result CLIENT **definition** LABELED-OBJECT SOURCE)

. Since the **definition** argument carried no information besides its
identity and the CHILDREN parameter was abused to pass the labeled
object, there was no way to pass any actual child parse results.

With this change, the above call becomes

  (make-expression-result CLIENT DEFINITION CHILDREN SOURCE)

where DEFINITION is an instance of the new definition class and
contains the labeled object. As a result, clients can specialize
methods to that class, obtain the labeled object and receive child
parse results in the CHILDREN parameter. The downsides of this change
are the increased protocol size due the exported LABELED-OBJECT reader
as well as the need to allocate DEFINITION and REFERENCE objects.