(Probably bad) Idea for labeled objects
The problem is that a client which defines a MAKE-LABELED-OBJECT
method does not see that object in

  (defmethod make-expression-result (client (result definition) ...)
    (labeled-object result))

but instead the wrapper object used by the parse-result module.

For clients that just want to stash the numeric label of the
definition/reference, we could probably just make the label available
in the DEFINITION and REFERENCE objects instead. But that would not
solve the general problem of the client not being able to access its
own labeled object.

The solution in this commit does allow the client to turn the opaque
labeled into the labeled object that the client made by introducing a
new return value for LABELED-OBJECT-STATE.

This is a bad solution. Hopefully, we can find a better one. One
alternative I tried is to store the inner labeled object as well as
well as the wrapper in the DEFINITION and REFERENCE instances and make
the LABELED-OBJECT reader return the inner labeled object but I
couldn't make that solution work so far.