Fix parse result for recovery from incomplete labeled object definition
Before this changes, one way to demonstrate the issue was reading an
input like

  #1=
  ;; a
  ;; b
  <eof>

and recovering from the error. The MAKE-EXPRESSION-RESULT method for
PARSE-RESULT-CLIENT was not prepared for a situation like the above in
which none of the child parse results of a labeled object definition
provides a "body" expression for the labeled object definition which
lead to an in correct parse result hierarchy.

With this change, only suitable parse result children are considered
as the "body" expression. In particular, if all existing children
resulted from skipped input, no special parse result will be made for
representing the labeled object definition. Instead, some other parse
result which represents the error recovery result will be returned.