New generic function FIND-PACKAGE
This simplifies the definition of the state-value protocol and
makes it possible for clients to use an alternate package system
without having to redefine INTERPRET-SYMBOL and the #+ #- macros.

I'm not sure how to provide a generic recovery for not finding a
keyword package. The extended package prefix code is also probably
a problem as it ought to signal an error.